LayoutTests/fast/canvas/text-globalAlpha-expected.txt

Ensure that globalAlpha is applied correctly to text.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS imgdata[3] == 127 || imgdata[3] == 128 is true
PASS successfullyParsed is true

TEST COMPLETE
